wip: induce unwinding when resetting a faulted instance
The current state of the repo is such that only the `fault_unwind` test is currently relevant:

```
cargo test -p lucet-runtime --test host fault_unwind -- --nocapture
```

Currently stuck figuring out how to set up the stack properly in order to return into the function
that panics. If I pad the stack with a zero word in order to keep it 16-byte aligned, the unwinding
runtime interprets that zero as a return address and fails. If I don't add the padding, later
instructions fault because of unaligned arguments.

We probably need to add a shim that uses `.cfi` directives in order to make the unwinding runtime
skip over the padding.